# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API для проекта Imperial",
        "title": "Imperial API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@imperial.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/v1/car-tags/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Удаляет тег по указанному ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Car Tags"
                ],
                "summary": "Удаление тега",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID тега",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Тег успешно удален",
                        "schema": {
                            "$ref": "#/definitions/car_tag.MessageResponse"
                        }
                    }
                }
            }
        },
...
        "/v1/cars/car-marks": {
            "get": {
                "description": "Возвращает список всех марок с поддержкой пагинации",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Car Marks"
                ],
                "summary": "Получение списка марок",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Смещение для пагинации",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Лимит для пагинации",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список марок",
                        "schema": {
                            "$ref": "#/definitions/car_mark.ListCarMarksResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Создает новую марку с указанным названием",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Car Marks"
                ],
                "summary": "Создание новой марки машины",
                "parameters": [
                    {
                        "description": "Данные для создания марки",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/car_mark.CreateCarMarkRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Марка успешно создана",
                        "schema": {
                            "$ref": "#/definitions/car_mark.CarMarkResponse"
                        }
                    }
                }
            }
        },
        "/v1/cars/car-marks/{id}": {
            "get": {
                "description": "Возвращает информацию о марке по указанному ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Car Marks"
                ],
                "summary": "Получение марки по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID марки",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация о марке",
                        "schema": {
                            "$ref": "#/definitions/car_mark.CarMarkResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Обновляет информацию о марке по указанному ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Car Marks"
                ],
                "summary": "Обновление марки",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID марки",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Новые данные для марки",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/car_mark.UpdateCarMarkRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Марка успешно обновлена",
                        "schema": {
                            "$ref": "#/definitions/car_mark.CarMarkResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Удаляет марку по указанному ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Car Marks"
                ],
                "summary": "Удаление марки",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID марки",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Марка успешно удалена",
                        "schema": {
                            "$ref": "#/definitions/car_mark.MessageResponse"
                        }
                    }
                }
            }
        },
        "/v1/cars/car-tags": {
            "get": {
                "description": "Возвращает список всех тегов с поддержкой пагинации",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Car Tags"
                ],
                "summary": "Получение списка тегов",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Смещение для пагинации",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Лимит для пагинации",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список тегов",
                        "schema": {
                            "$ref": "#/definitions/car_tag.ListCarTagsResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Создает новый тег с указанным названием",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Car Tags"
                ],
                "summary": "Создание нового тега для машины",
                "parameters": [
                    {
                        "description": "Данные для создания тега",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/car_tag.CreateCarTagRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Тег успешно создан",
                        "schema": {
                            "$ref": "#/definitions/car_tag.CarTagResponse"
                        }
                    }
                }
            }
        },
        "/v1/cars/car-tags/{id}": {
            "get": {
                "description": "Возвращает информацию о теге по указанному ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Car Tags"
                ],
                "summary": "Получение тега по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID тега",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация о теге",
                        "schema": {
                            "$ref": "#/definitions/car_tag.CarTagResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Обновляет информацию о теге по указанному ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Car Tags"
                ],
                "summary": "Обновление тега",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID тега",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Новые данные для тега",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/car_tag.UpdateCarTagRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Тег успешно обновлен",
                        "schema": {
                            "$ref": "#/definitions/car_tag.CarTagResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "car_mark.CarMarkResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "car_mark.CreateCarMarkRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "example": "Toyota"
                }
            }
        },
        "car_mark.ListCarMarksResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entities.CarMark"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "car_mark.MessageResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "car_mark.UpdateCarMarkRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "example": "Toyota"
                }
            }
        },
        "car_tag.CarTagResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "car_tag.CreateCarTagRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "example": "Sedan"
                }
            }
        },
        "car_tag.ListCarTagsResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entities.CarTag"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "car_tag.MessageResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "car_tag.UpdateCarTagRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "example": "Sedan"
                }
            }
        },
...
        "entities.CarMark": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "entities.CarTag": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT token must be passed with `Bearer ` prefix. Example: \"Bearer eyJhbGciOiJIUzI1NiI...\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}